requirement.txt
Pillow
transformers
torch